Move SemanticDocumentationArticle to Models namespaces
This is a model, not service
