+++ packages/framework/resources/views/components/docs/documentation-article.blade.php
@@ -1,5 +1,5 @@
 @props([
-    /** @var \Hyde\Framework\Services\SemanticDocumentationArticle */
+    /** @var \Hyde\Support\Models\SemanticDocumentationArticle */
     'document'
 ])
